Move configure_search method into indexer
